Renaming 'com.qubitpi' to 'org.qubitpi'
