+++ .github/workflows/ci-cd.yaml
@@ -49,19 +49,9 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actionshub/markdownlint@main
 
-  linkChecker:
-    name: Link Check
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Link Checker
-        uses: lycheeverse/lychee-action@v1.9.0
-        with:
-          fail: true
-
   tests:
     name: Unit & Integration Tests
-    needs: [yaml-lint, markdown-lint, linkChecker]
+    needs: [yaml-lint, markdown-lint]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
