+++ src/test/groovy/org/qubitpi/fastws/JettyServerFactorySpec.groovy
@@ -13,19 +13,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.fastws.web.endpoints
+package org.qubitpi.fastws
 
-import com.qubitpi.fastws.JettyServerFactory
-import com.qubitpi.fastws.application.ResourceConfig
+import org.qubitpi.fastws.resource.TestEndpoint
 
 import org.eclipse.jetty.server.Server
+import org.glassfish.jersey.server.ResourceConfig
 
 import io.restassured.RestAssured
+import jakarta.inject.Inject
+import jakarta.ws.rs.ApplicationPath
 import spock.lang.Specification
 
-class HealthcheckServletITSpec extends Specification {
+class JettyServerFactorySpec extends Specification {
 
     static final int PORT = 8080
+    static final String ENDPOINT_RESOURCE_PACKAGE = TestEndpoint.class.getPackage().getName()
+
+    /**
+     * DI constructor.
+     * <p>
+     * CAUTION: the {@code @ApplicationPath("v1")} is not taking effects. See {@link JettyServerFactory} for more
+     * details.
+     */
+    @ApplicationPath("v1")
+    class TestResourceConfig extends ResourceConfig {
+
+        @Inject
+        TestResourceConfig() {
+            packages(ENDPOINT_RESOURCE_PACKAGE)
+        }
+    }
 
     def setupSpec() {
         RestAssured.baseURI = "http://localhost"
@@ -33,17 +51,21 @@ class HealthcheckServletITSpec extends Specification {
         RestAssured.basePath = "/v1"
     }
 
-    def "Healthchecking endpoints returns 200"() {
+    def "Factory produces Jersey-Jetty applications"() {
         setup:
-        Server server = JettyServerFactory.newInstance(PORT, "/v1/*", new ResourceConfig())
+        Server server = JettyServerFactory.newInstance(PORT, "/v1/*", new TestResourceConfig())
         server.start()
 
         expect:
-        RestAssured.given()
+        RestAssured
                 .when()
-                .get("/healthcheck")
+                .get("/v1/example/test")
                 .then()
                 .statusCode(200)
+
+        RestAssured
+                .when()
+                .get("/v1/example/test").asString() == "SUCCESS"
 
         cleanup:
         server.stop()
