+++ src/test/groovy/org/qubitpi/fastws/web/filters/CorsFilterSpec.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.fastws.web.filters
+package org.qubitpi.fastws.web.filters
 
 import jakarta.ws.rs.container.ContainerRequestContext
 import jakarta.ws.rs.container.ContainerResponseContext
